refactor(docker): create ownership column
