+++ app/react/docker/containers/ListView/ContainersDatatable/columns/index.tsx
@@ -1,12 +1,14 @@
 import _ from 'lodash';
 import { useMemo } from 'react';
 
+import { createOwnershipColumn } from '@/react/docker/components/datatable-helpers/createOwnershipColumn';
+import { DockerContainer } from '@/react/docker/containers/types';
+
 import { created } from './created';
 import { host } from './host';
 import { image } from './image';
 import { ip } from './ip';
 import { name } from './name';
-import { ownership } from './ownership';
 import { ports } from './ports';
 import { quickActions } from './quick-actions';
 import { stack } from './stack';
@@ -30,7 +32,7 @@ export function useColumns(
         isHostColumnVisible && host,
         isGPUsColumnVisible && gpus,
         ports,
-        ownership,
+        createOwnershipColumn<DockerContainer>(),
       ]),
     [isHostColumnVisible, isGPUsColumnVisible]
   );
